requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
boto3
pydantic
python-magic
python-magic-bin
apscheduler
shippo
PyJWT
passlib
python-jose
bcrypt
pytest